Update creatures.json
Added all German names of the creatures as they should be and how they are in the official translation of the game. 
Most of the creatures got the english name, and should be getting those in the Dododex App as well, I know that the person wanted to translate them in German, but most of them are not even the same as in the german translation. 

~Miyu Lin Xiahou